Added ServerBrowser tab with a Mod-Sync feature.
To host a server with Mod-Sync capabilities you must use a modified
mtndew.dll (ask in the ED IRC for help with this). You do not need the
.dll as a client.

When a server is selected, if it "requires" mods to be installed, the
user will have a list of mods and a "Mod-Sync" button will appear.
Clicking this button will kill all running ED processes, download each
mod that needs to be download, and then install each mod that needs to
be installed.