Added some #if RELEASE/DEBUG stuff to help with debugging ServerBrowser and ModSync
Added an #if RELEASE around the portions of code that check if ED
version matches a server for the QuickMatch and Connect buttons.

Added an #if DEBUG around some new code inside the
functions_UpdateServerList.cs file, which adds a fake server to the
servers list (can easily edit the values to test a variety of things).